moves availability componets to thier own folder
